# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Email Converter API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ConversionJobFileType =
  (typeof ConversionJobFileType)[keyof typeof ConversionJobFileType];

export const ConversionJobFileType = {
  eml: "eml",
  mbox: "mbox",
} as const;

export type ConversionJobStatus =
  (typeof ConversionJobStatus)[keyof typeof ConversionJobStatus];

export const ConversionJobStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

export interface ConversionJob {
  id: number;
  originalFilename: string;
  fileType: ConversionJobFileType;
  status: ConversionJobStatus;
  totalEmails: number;
  processedEmails: number;
  createdAt: string;
  updatedAt: string;
  errorMessage?: string | null;
}

export interface ConvertedEmail {
  id: number;
  jobId: number;
  generatedFilename: string;
  subject: string;
  sender: string;
  recipients: string;
  emailDate?: string | null;
  bodyText: string;
  hasAttachments: boolean;
  attachmentCount: number;
  attachmentNames?: string | null;
  createdAt: string;
}

export interface JobListResponse {
  jobs: ConversionJob[];
  total: number;
}

export interface JobDetailResponse {
  job: ConversionJob;
  emails: ConvertedEmail[];
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
  details?: string | null;
}

export type UploadEmailFileBody = {
  /** The .eml or .mbox file to convert */
  file: Blob;
};

export type ListJobsParams = {
  limit?: number;
  offset?: number;
};