# lib/api-zod/src/generated/types/conversionJob.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Email Converter API
 * OpenAPI spec version: 0.1.0
 */
import type { ConversionJobFileType } from "./conversionJobFileType";
import type { ConversionJobStatus } from "./conversionJobStatus";

export interface ConversionJob {
  id: number;
  originalFilename: string;
  fileType: ConversionJobFileType;
  status: ConversionJobStatus;
  totalEmails: number;
  processedEmails: number;
  createdAt: Date;
  updatedAt: Date;
  errorMessage?: string | null;
}