# lib/api-zod/src/generated/types/jobDetailResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Email Converter API
 * OpenAPI spec version: 0.1.0
 */
import type { ConversionJob } from "./conversionJob";
import type { ConvertedEmail } from "./convertedEmail";

export interface JobDetailResponse {
  job: ConversionJob;
  emails: ConvertedEmail[];
}