# lib/api-zod/src/generated/types/jobListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Email Converter API
 * OpenAPI spec version: 0.1.0
 */
import type { ConversionJob } from "./conversionJob";

export interface JobListResponse {
  jobs: ConversionJob[];
  total: number;
}